face_index/requirements.txt
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.41.2
numpy==1.24.2
fastapi==0.95.0
uvicorn==0.21.1
flake8==5.0.4
requests==2.28.2
python-logstash-async==2.5.0
python-multipart==0.0.6
talisman-dm==0.7.1
werkzeug==2.2.1
faiss-cpu==1.7.4
psycopg2-binary==2.9.6